feat: restructure as Cargo workspace for multi-language bindings
## Summary
- Convert single-crate project to Cargo workspace structure
- Separate core Rust library from WASM bindings
- Enable future language bindings (Python via Maturin)
- Fix npm package naming issues with wasm-pack

## Workspace Structure
- `pubmed-client/` - Core Rust library (pubmed-client-rs)
- `pubmed-client-wasm/` - WASM bindings for npm (pubmed-client-wasm)
- Shared dependencies and test data

## Key Changes
- Root `Cargo.toml` now defines workspace
- WASM package correctly generates `pubmed-client-wasm` name for npm
- All npm/TypeScript files moved to WASM package directory
- Updated CLAUDE.md with comprehensive workspace documentation
- Fixed Vitest configuration and test imports
- All tests passing: 209 unit tests + 16 WASM tests
