+++ Cargo.toml
@@ -1,129 +1,19 @@
-[package]
-name = "pubmed-client-rs"
-version = "0.1.0"
-autotests = false
-edition = "2024"
+[workspace]
+members = ["pubmed-client", "pubmed-client-wasm"]
 resolver = "2"
 
-[lib]
-crate-type = ["cdylib", "rlib"]
+[workspace.package]
+version = "0.1.0"
+authors = ["illumination-k"]
+edition = "2021"
+license = "MIT"
+repository = "https://github.com/illumination-k/pubmed-client-rs"
 
-[dependencies]
+[workspace.dependencies]
 anyhow = "1.0"
 quick-xml = { version = "0.36", features = ["serialize"] }
-rand = "0.8"
-regex = "1.10"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 thiserror = "1.0"
 tracing = "0.1"
 urlencoding = "2.1"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-reqwest = { version = "0.11", features = ["json"] }
-tokio = { version = "1.0", features = ["full"] }
-tokio-retry = "0.3"
-tokio-util = "0.7"
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-reqwest = { version = "0.11", features = ["json"], default-features = false }
-tokio = { version = "1.0", features = ["macros", "rt", "time"], default-features = false }
-getrandom = { version = "0.2", features = ["js"] }
-# WASM bindings - controlled by 'wasm' feature
-wasm-bindgen = { version = "0.2", optional = true }
-wasm-bindgen-futures = { version = "0.4", optional = true }
-js-sys = { version = "0.3", optional = true }
-web-sys = { version = "0.3", optional = true, features = ["console"] }
-serde-wasm-bindgen = { version = "0.6", optional = true }
-tsify = { version = "0.4", optional = true }
-console_error_panic_hook = { version = "0.1", optional = true }
-wee_alloc = { version = "0.4", optional = true }
-
-[features]
-default = []
-integration-tests = []
-wasm = [
-  "wasm-bindgen",
-  "wasm-bindgen-futures",
-  "js-sys",
-  "web-sys",
-  "serde-wasm-bindgen",
-  "tsify",
-  "console_error_panic_hook",
-  "wee_alloc",
-]
-
-[dev-dependencies]
-rstest = "0.18"
-tokio-test = "0.4"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-tracing-test = "0.2"
-wiremock = "0.6"
-
-# Integration tests
-# Comprehensive test suites
-[[test]]
-name = "comprehensive_pmc_tests"
-path = "tests/integration/comprehensive_pmc_tests.rs"
-
-[[test]]
-name = "comprehensive_pubmed_tests"
-path = "tests/integration/comprehensive_pubmed_tests.rs"
-
-[[test]]
-name = "comprehensive_einfo_tests"
-path = "tests/integration/comprehensive_einfo_tests.rs"
-
-[[test]]
-name = "comprehensive_elink_tests"
-path = "tests/integration/comprehensive_elink_tests.rs"
-
-# Feature-specific tests
-[[test]]
-name = "markdown_tests"
-path = "tests/integration/markdown_tests.rs"
-
-[[test]]
-name = "pmc_xml_tests"
-path = "tests/integration/pmc_xml_tests.rs"
-
-[[test]]
-name = "test_abstract_parsing"
-path = "tests/integration/test_abstract_parsing.rs"
-
-[[test]]
-name = "test_einfo_integration"
-path = "tests/integration/test_einfo_integration.rs"
-
-[[test]]
-name = "test_elink_integration"
-path = "tests/integration/test_elink_integration.rs"
-
-[[test]]
-name = "test_integration_abstract"
-path = "tests/integration/test_integration_abstract.rs"
-
-[[test]]
-name = "test_mesh_integration"
-path = "tests/integration/test_mesh_integration.rs"
-
-[[test]]
-name = "test_rate_limiting_mocked"
-path = "tests/integration/test_rate_limiting_mocked.rs"
-
-# Real API tests (opt-in)
-[[test]]
-name = "test_real_api_rate_limiting"
-path = "tests/integration/test_real_api_rate_limiting.rs"
-
-[[test]]
-name = "pubmed_api_tests"
-path = "tests/integration/pubmed_api_tests.rs"
-
-[[test]]
-name = "pmc_api_tests"
-path = "tests/integration/pmc_api_tests.rs"
-
-[[test]]
-name = "error_handling_tests"
-path = "tests/integration/error_handling_tests.rs"
