+++ pubmed-client-wasm/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "pubmed-client-wasm"
+version.workspace = true
+authors.workspace = true
+categories = ["api-bindings", "wasm", "web-programming"]
+edition.workspace = true
+keywords = ["pubmed", "wasm", "webassembly", "biomedical", "api"]
+license.workspace = true
+readme = "README.md"
+repository.workspace = true
+description = "WebAssembly bindings for the PubMed client library"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+pubmed-client-rs = { path = "../pubmed-client" }
+
+# WASM-specific dependencies
+console_error_panic_hook = "0.1"
+js-sys = "0.3"
+serde-wasm-bindgen = "0.6"
+tsify = "0.4"
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
+web-sys = { version = "0.3", features = ["console"] }
+wee_alloc = "0.4"
+
+# Required for WASM target
+getrandom = { version = "0.2", features = ["js"] }
+
+# Re-export some core dependencies for WASM use
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+
+[features]
+default = []
