+++ pubmed-client-wasm/package.json
@@ -2,20 +2,21 @@
   "name": "pubmed-client-wasm",
   "version": "0.1.0",
   "description": "WebAssembly bindings for the PubMed client library",
-  "main": "pkg/pubmed_client_rs.js",
-  "types": "pkg/pubmed_client_rs.d.ts",
+  "main": "pkg/pubmed_client_wasm.js",
+  "types": "pkg/pubmed_client_wasm.d.ts",
   "files": [
     "pkg/"
   ],
   "scripts": {
-    "build": "wasm-pack build --target nodejs --features wasm",
-    "build:web": "wasm-pack build --target web --features wasm",
-    "build:bundler": "wasm-pack build --target bundler --features wasm",
+    "build": "wasm-pack build --target nodejs",
+    "build:web": "wasm-pack build --target web",
+    "build:bundler": "wasm-pack build --target bundler",
     "build:all": "pnpm run build && pnpm run build:web && pnpm run build:bundler",
+    "publish": "wasm-pack publish --access public",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
-    "test:wasm": "wasm-pack test --node --features wasm",
+    "test:wasm": "wasm-pack test --node",
     "typecheck": "tsc --noEmit",
     "lint": "biome lint .",
     "format": "biome format --write .",
@@ -24,7 +25,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/illumination-k/pubmed-client-rs.git"
+    "url": "git+https://github.com/illumination-k/pubmed-client-rs.git"
   },
   "keywords": [
     "pubmed",
@@ -35,8 +36,8 @@
     "webassembly",
     "rust"
   ],
-  "author": "Your Name",
-  "license": "MIT OR Apache-2.0",
+  "author": "illumination-k",
+  "license": "MIT",
   "devDependencies": {
     "@biomejs/biome": "^2.0.6",
     "@types/node": "^20.0.0",
