+++ pubmed-client/src/config.rs
@@ -434,11 +434,9 @@ mod tests {
             config.effective_base_url(),
             "https://eutils.ncbi.nlm.nih.gov/entrez/eutils"
         );
-        assert!(
-            config
-                .effective_user_agent()
-                .starts_with("pubmed-client-rs/")
-        );
+        assert!(config
+            .effective_user_agent()
+            .starts_with("pubmed-client-rs/"));
         assert_eq!(config.effective_tool(), "pubmed-client-rs");
     }
