+++ pubmed-client/src/error.rs
