+++ pubmed-client/src/pmc/client.rs
