+++ pubmed-client/src/pubmed/models.rs
