+++ pubmed-client/src/pubmed/query/search.rs
