+++ tests/integration/comprehensive_pmc_tests.rs
@@ -4,7 +4,7 @@ use tracing::{debug, info, warn};
 use pubmed_client_rs::pmc::PmcXmlParser;
 
 mod common;
-use common::{PmcXmlTestCase, pmc_xml_test_cases};
+use common::{pmc_xml_test_cases, PmcXmlTestCase};
 
 /// 全XMLファイルを返すフィクスチャ
 #[fixture]
