+++ tests/integration/pmc_api_tests.rs
@@ -30,7 +30,7 @@ mod integration_tests {
 
     // Import test utilities
     use crate::common::integration_test_utils::{
-        TEST_PMCIDS, create_test_pmc_client, create_test_pubmed_client, should_run_real_api_tests,
+        create_test_pmc_client, create_test_pubmed_client, should_run_real_api_tests, TEST_PMCIDS,
     };
 
     /// Test fetching PMC full-text articles
